+++ fennel/sources/sources.py
@@ -40,8 +40,8 @@ def source(
     conn: DataConnector,
     every: Optional[Duration] = None,
     since: Optional[datetime] = None,
-    disorder: Optional[Duration] = None,
-    cdc: Optional[str] = None,
+    disorder: Duration = DEFAULT_DISORDER,
+    cdc: str = DEFAULT_CDC,
     tier: Optional[Union[str, List[str]]] = None,
     preproc: Optional[Dict[str, PreProcValue]] = None,
 ) -> Callable[[T], Any]:
@@ -58,8 +58,8 @@ def source(
 
     def decorator(dataset_cls: T):
         conn.every = every if every is not None else DEFAULT_EVERY
-        conn.disorder = disorder if disorder is not None else DEFAULT_DISORDER
-        conn.cdc = cdc if cdc is not None else DEFAULT_CDC
+        conn.disorder = disorder
+        conn.cdc = cdc
         conn.since = since
         conn.tiers = TierSelector(tier)
         conn.pre_proc = preproc